+++ src/grz_cli/progress/progress_logging.py
@@ -11,7 +11,7 @@ from pathlib import Path
 
 from grz_pydantic_models.submission.metadata.v1 import File as SubmissionFileMetadata
 
-from .file_operations import read_multiple_json
+from ..utils.io import read_multiple_json
 from .states import State
 
 Index = tuple[str, float, int]
