+++ src/grz_cli/workers/upload.py
@@ -17,8 +17,7 @@ from tqdm.auto import tqdm
 
 from ..constants import TQDM_SMOOTHING
 from ..models.config import ConfigModel
-from ..progress_logging import FileProgressLogger
-from ..states import UploadState
+from ..progress import FileProgressLogger, UploadState
 from ..transfer import init_s3_client
 
 MULTIPART_THRESHOLD = 8 * 1024**2  # 8MiB, boto3 default, largely irrelevant
