+++ tests/test_bam_validation.py
@@ -2,7 +2,7 @@
 
 import importlib.resources
 
-from grz_cli.bam_validation import validate_bam
+from grz_cli.validation import validate_bam
 
 from . import resources
