+++ tests/test_fastq_validation.py
@@ -1,6 +1,6 @@
 """Tests for the fastq_validation module."""
 
-from grz_cli.fastq_validation import (
+from grz_cli.validation import (
     validate_paired_end_reads,
     validate_single_end_reads,
 )
