+++ tests/test_file_progress_logging.py
@@ -7,8 +7,8 @@ from pathlib import Path
 import pytest
 from grz_pydantic_models.submission.metadata.v1 import File as SubmissionFileMetadata
 
-from grz_cli.file_operations import read_multiple_json
-from grz_cli.progress_logging import FileProgressLogger
+from grz_cli.progress import FileProgressLogger
+from grz_cli.utils.io import read_multiple_json
 
 
 @pytest.fixture(scope="function")
